+++ src/main/java/uk/gov/companieshouse/api/accounts/validation/CurrentPeriodValidator.java
@@ -34,11 +34,11 @@ public class CurrentPeriodValidator extends BaseValidator {
         FixedAssets fixedAssets = currentPeriod.getBalanceSheet().getFixedAssets();
         if (fixedAssets != null) {
 
-            Integer tangible = fixedAssets.getTangible();
-            Integer fixedAssetsTotal = fixedAssets.getTotalFixedAssets();
+            Long tangible = fixedAssets.getTangible();
+            Long fixedAssetsTotal = fixedAssets.getTotalFixedAssets();
 
             // Will calculate the total of all fixedAssets fields as they are added to the balance sheet
-            Integer calculatedTotal = tangible;
+            Long calculatedTotal = tangible;
 
             validateAggregateTotal(fixedAssetsTotal, calculatedTotal, TOTAL_PATH, errors);
         }
